:recycle: refactor: update fixtures file
