+++ texttest_fixtures.go
@@ -11,6 +11,18 @@ import (
 func main() {
 	fmt.Println("OMGHAI!")
 
+	var items = []*gildedrose.Item{
+		{"+5 Dexterity Vest", 10, 20},
+		{"Aged Brie", 2, 0},
+		{"Elixir of the Mongoose", 5, 7},
+		{"Sulfuras, Hand of Ragnaros", 0, 80},
+		{"Sulfuras, Hand of Ragnaros", -1, 80},
+		{"Backstage passes to a TAFKAL80ETC concert", 15, 20},
+		{"Backstage passes to a TAFKAL80ETC concert", 10, 49},
+		{"Backstage passes to a TAFKAL80ETC concert", 5, 49},
+		{"Conjured Mana Cake", 3, 6}, // <-- :O
+	}
+
 	days := 2
 	var err error
 	if len(os.Args) > 1 {
